document, switch to int, add more tests, style

+++ src/relay/op/make_op.h
@@ -62,7 +62,7 @@ Expr MakeRepeat(Expr data, int repeats, int axis);
 
 Expr MakeReshape(Expr data, Array<Integer> newshape);
 
-Expr MakeReshapeLike(Expr lhs, Expr rhs, int64_t lhs_begin, Integer lhs_end, int64_t rhs_begin,
+Expr MakeReshapeLike(Expr lhs, Expr rhs, int lhs_begin, Integer lhs_end, int rhs_begin,
                      Integer rhs_end);
 
 Expr MakeSplit(Expr data, ObjectRef indices_or_sections, int axis);
